Move useManagePosition to Mange page
