+++ v3/ui/src/pages/Manage/ManageActions.tsx
@@ -1,5 +1,5 @@
 import { ArrowDownIcon, ArrowUpIcon } from '@chakra-ui/icons';
-import { Box, Button, Flex, Text } from '@chakra-ui/react';
+import { Box, Button, Flex } from '@chakra-ui/react';
 import { BorderBox } from '@snx-v3/BorderBox';
 import { BorrowIcon, DollarCircle } from '@snx-v3/icons';
 import { ManagePositionContext } from '@snx-v3/ManagePositionContext';
@@ -11,7 +11,7 @@ import { wei } from '@synthetixio/wei';
 import { FC, FormEvent, PropsWithChildren, useCallback, useContext } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import { Borrow } from './Borrow';
-import { useManagePosition } from '../AccountPositionPage/useManagePosition';
+import { useManagePosition } from './useManagePosition';
 import { Repay } from './Repay';
 import { Withdraw } from './Withdraw';
 import { Deposit } from './Deposit';
